requirements.txt
celery==3.1.17
django-bulbs==0.4.8
uwsgi==2.0.9